+++ config.yaml.example
@@ -1,0 +1,2 @@
+OPENAI_API_KEY: "ENTER YOUR OPENAI KEY"
+SERPER_DEV_API_KEY: "ENTER YOUR SERPER DEV KEY"
